Move WavLM attention to a separate file
